grpc_client/requirements.txt
grpcio
grpcio-tools
protobuf
fastapi[standard]
uvicorn